refactor: seperate constant to new file not done yet
