+++ index.html
@@ -121,6 +121,8 @@
         </div>
     </form>
     </divv>
-    <script src="public/js/app.js"></script>
+    <script type="module" src="public/js/app.js"></script>
+    <!-- <script type="module" src="public/js/constants.js"></script> -->
+
   </body>
 </html>
