+++ src/constants.ts
@@ -1,0 +1,23 @@
+const COLOR_TYPES: ColorTypes = {
+  GRASS: "rgb(99, 194, 99)",
+  POISON: "rgb(150, 29, 150)",
+  BUG: "rgb(31, 147, 31)",
+  NORMAL: "rgb(147, 146, 146);",
+  FLYING: "rgb(73, 73, 207)",
+  WATER: "rgb(108, 177, 220)",
+  FIRE: "rgb(186, 63, 63)",
+  ELECTRIC: "rgb(193, 193, 107)",
+};
+
+const POKE_TYPES: PokeTypes = {
+  GRASS: "grass",
+  POISON: "poison",
+  BUG: "bug",
+  NORMAL: "normal",
+  FLYING: "flying",
+  WATER: "water",
+  FIRE: "fire",
+  ELECTRIC: "electric",
+};
+
+export { POKE_TYPES, COLOR_TYPES };
